backend/OngsDoBem/requirements.txt
asgiref==3.5.0
cryptocode==0.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
pycryptodomex==3.14.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
